+++ api/src/adapters/storage/__init__.py
@@ -1,2 +1,1 @@
-from .base import Storage
 from .exoscale import Exoscale
